@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
 
 class Validity {
-  final int validThruMonth;
-  final int validThruYear;
-  final int validFromMonth;
-  final int validFromYear;
+  final int? validThruMonth;
+  final int? validThruYear;
+  final int? validFromMonth;
+  final int? validFromYear;
 
   Validity({
-    @required this.validThruMonth,
-    @required this.validThruYear,
+    required this.validThruMonth,
+    required this.validThruYear,
     this.validFromMonth,
     this.validFromYear,
   }) {
